fixture consolidation
